Add explicit dependencies
Without the explicit dependencies, the rancher_host resource is created from
the start, before the provisioner can even be started. This has the consequence
that its timeout can expire if the provisioner is too long to start (due to
it's dependencies not being ready) or too run. By adding explicit dependencies,
this resource will be created only when the provisioner is successful which
will solve the problem.